Add files via upload
MOviesWatcher 🎬
MOviesWatcher is a web-based application that allows users to search for movies and explore popular titles. It uses The Movie Database (TMDb) API to retrieve data and display it in an intuitive and responsive interface.

Features
Popular Movies Display: Automatically displays popular movies on the homepage.
Search Functionality: Allows users to search for any movie by title.
Movie Ratings: Displays color-coded ratings for movies (Green for high ratings, Orange for average, and Red for low ratings).
Overview Pop-Up: Shows a brief overview of each movie on hover for quick access to details.